Separate out the types, since importing WorkerRequestTypes from the frontend page (of the worker) seemed to cause the 'self' error. May still need to fix that import or revert using enums.
